added logging, error checking, and general cleanup

+++ gmail_to_db/gmail_to_db.py
@@ -2,10 +2,18 @@ from __future__ import with_statement
 import base64
 import email.parser
 import imaplib
+import logging
 import os
 import re
+import sys
 
-save_directory = os.getcwd() # directory where to save attachments (default: current)
+# Set up logging
+logging.basicConfig(filename='gmail_to_db.log', level=logging.DEBUG,
+                    format='%(asctime)s %(levelname)s %(message)s')
+logging.debug('Starting gmail_to_db pull script')
+
+# Global (constant) variables
+save_directory = os.getcwd()
 user = 'daphotline'
 password = base64.b64decode('ZmViMTEyMDEx')
 
@@ -17,23 +25,43 @@ gmail.select("[Gmail]/All Mail")
 
 # Download the email ids
 ok_response, item_string = gmail.search(None, "ALL")
-assert(ok_response == 'OK')  # TODO(topher): do something on error response
+if (ok_response != 'OK'):
+    logging.error('gmail.search returned error: {err}'. format(err=ok_response))
+    sys.exit()
 email_ids = map(int, item_string[0].split())
 
-for email_id in email_ids:
-    print email_id
-
-    # fetching the mail, "`(RFC822)`" means "get the whole stuff", but you can
-    # ask for headers only, etc
+# Check each email that we an id for to see if it's a message we want
+for email_id in email_ids[-1:]:
+    logging.debug("processing email_id: {id}".format(id=email_id))
+    # fetching the mail, "`(RFC822)`" means "get the whole stuff"
     ok_response, data = gmail.fetch(email_id, "(RFC822)") 
-    assert(ok_response == 'OK')  # TODO(topher): do something on error response
+    if (ok_response != 'OK'):
+        logging.error('gmail.fetch returned error: {err}'.
+                     format(err=ok_response))
+        sys.exit()
 
-    email_body = data[0][1] # getting the mail content
+    # getting the mail content
+    email_body = data[0][1]
     mail = email.message_from_string(email_body)
-    for part in  mail.walk():
-        print part.get('Content-Type')
+
+    # Make sure we're looking at an email from phonepeople.com
+    if (not re.search('phonepeople.com', mail["From"])):
+        logging.debug("The email isn't from phonepeople.com- skipping")
+        sys.exit()
+
+    # Check each part of the email to find (and save) the .wav file
+    for part in mail.walk():
         if re.match('audio/wav', part.get('Content-Type')):
-            print "found it!!!"
+            logging.info("We've found a .wav from phonepeople!")
             abs_filename = os.path.join(save_directory, part.get_filename())
+
+            # Make sure it isn't already a file
+            if os.path.isfile(abs_filename):
+                logging.error("The file {filename} already exists".
+                              format(filename=abs_filename))
+                continue
+
+            # Save the file
+            # TODO(topher): make this add to database instead of disk
             with open(abs_filename, 'wb') as file:
                 file.write(part.get_payload(decode=True))
